% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.fit.preparation.R
\name{length.fit.preparation}
\alias{length.fit.preparation}
\title{Read in and process length.fit file}
\usage{
\method{length}{fit.preparation}(filename)
}
\arguments{
\item{filname}{The filename of the length.fit file, including the location.}
}
\description{
Hoovers out the observed ann predicted composite (i.e. summed over time) catches at length from the length.fit file.
It should also work with the weight.fit file but hasn't yet been tested.
Does not handle multispecies.
}
