% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.r
\docType{methods}
\name{jitter}
\alias{jitter}
\title{jitter}
\usage{
jitter(par, sd, seed)
}
\arguments{
\item{object:}{An object of class MFCLPar}
}
\value{
An object of class MFCLPar
}
\description{
Function to take a par file and change the values within based on an input SD to allow to perform a jittering analysis for a model in MFCL
This probably only works for a single species model and has not been coded to be useable by the multispecies model
}
\details{
This function is still preliminary and in development. Currently things that are not completed includin those below
new orthogonal coefficients
extra fisheries pars except for 1,2,4
length based selectivity parameters
selectivity deviates
mean natural mortality and deviates/ functional forms
growth curve deviates ??? different from von Bertalannfy growth deviates which is accounted for
Seasonal griwth parameter
age dependent movement coefficients
nonlinear movement coefficients

Probably others that I didn't notice or don't have any idea about such as the Lagrangian
}
