% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.age.dist.r
\name{plot.age.dist}
\alias{plot.age.dist}
\title{Plot the distribution an age class by region and quarter}
\usage{
plot.age.dist(
  rep,
  ages,
  year_range = (as.numeric(range(rep)["minyear"])):as.numeric(range(rep)["maxyear"]),
  plot_type = "violin",
  overlay_data = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{rep}{a MFCLRep object}

\item{ages}{The age class that you want to plot the distribution of values across}

\item{year_range}{The year range to plot the distribution of the recruitment over. Default is all years}

\item{plot_type}{can be "violin" for a violin plot or "box" for a boxplot. Default is violin plot}

\item{overlay_data}{Do you want to overlay the original data on the distribution? TRUE or FALSE (default).}

\item{palette.func}{a function to determine the colours of each area.  It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the output will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguements to the palette function. Use the aurgument all.model.colours to ensure consistency of model colours when plottin a subset of models.}
}
\description{
Plot the distribution of proportions of a chosen age class across year between quarters and region from a single model.
}
