% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.biomass.contributions.r
\name{plot.biomass.contributions}
\alias{plot.biomass.contributions}
\title{Plot the contributions to biomass from each region
Original code from Rob Scott's FLR4MFCL package and refreshed for diags4mfcl}
\usage{
plot.biomass.contributions(
  par,
  rep,
  Fmult = 0,
  proportion = TRUE,
  cols = NULL,
  save.dir,
  save.name
)
}
\arguments{
\item{par}{MFCLPar object}

\item{rep}{MFCLRep object}

\item{Fmult}{A multiplier on the terminal year fishing mortality}

\item{proportion}{TRUE or FALSE. Plot raw biomass (shows scale between region) or plot proportion of biomass by source region}

\item{cols}{A vector of colors for each region. If NULL then the default will be a vibrant rainbow.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the contributions to biomass from each region
Original code from Rob Scott's FLR4MFCL package and refreshed for diags4mfcl
}
