% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.biomass.r
\name{plot.biomass}
\alias{plot.biomass}
\title{Compare the estimated biomass across different models.}
\usage{
plot.biomass(
  rep.list,
  rep.names,
  agg.years = TRUE,
  agg.regions = TRUE,
  biomass.type = "SSB",
  palette.cols = c("royalblue3", "deepskyblue1", "gold", "orange1", "indianred1",
    "firebrick2", "#AC2020"),
  save.dir,
  save.name
)
}
\arguments{
\item{rep.list}{A list of MFCLRep objects or a single MFCLRep object. The reference model should be listed first.}

\item{rep.names}{A vector of character strings naming the models for plotting purposes}

\item{agg.years}{TRUE or FALSE. Should model outputs be aggregated to an annual time step.}

\item{agg.regions}{TRUE or FALSE. Should model outputs be aggregated across all regions are kept separate.}

\item{biomass.type}{Character string denoting the type of biomass plotted, 'SSB' or 'Total'}

\item{palette.cols}{A vector of character strings giving the colors to form a palette for differentiating between models. If wishing to use the exact colors give a vector of (length(rep.list) - 1) colors as the reference model is black by default.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Compare the estimated biomass across different models.
}
