% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.comparisons.r
\name{plot.comparisons}
\alias{plot.comparisons}
\title{Wrapper function for plotting model comparisons}
\usage{
plot.comparisons(
  plot.name,
  rep.list,
  rep.names,
  agg.years = TRUE,
  agg.regions = TRUE,
  sel.basis = "AGE",
  biomass.type = "SSB",
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{rep.list}{A list of MFCLRep objects. The reference model should be listed first.}

\item{rep.names}{A vector of character strings naming the models for plotting purposes}

\item{agg.years}{TRUE or FALSE. Should model outputs be aggregated to an annual time step.}

\item{agg.regions}{TRUE or FALSE. Should model outputs be aggregated across all regions are kept separate.}

\item{sel.basis}{A character string indicating if selectivity at age ('AGE') or length ('Length') should be plotted}

\item{biomass.type}{Character string denoting the type of biomass plotted, 'SSB' or 'Total'}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.colours to ensure consistency of model colours when plotting a subset of models.}
}
\description{
Wrapper function for plotting model comparisons
}
