% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.everything.r
\name{plot.everything}
\alias{plot.everything}
\title{Create all of the diagnostic plots from a given directory}
\usage{
plot.everything(rundir, spp = "skj", par, fdescloc = NULL, outdir = NULL, ...)
}
\arguments{
\item{rundir}{A directory string that tells where to get the files}

\item{spp}{The species abbreviation that comes before}

\item{par}{optional name of par file you you want plotted otherwise finds and uses largest number value}

\item{fdescloc}{optional location of fdesc.txt file that contains information for labels of figures if not included within the folder. the fdesc.txt is a file with 7 columns and an entry for each fishery in the model. The columns should be named 1. num 2. gear_long  3. method  4. code  5. gear 6. flag 7. region. This file creates all of the labels for the fisheries.}

\item{outdir}{An optional directory path that tells where the figures are to be saved. The function will automatically create a directory called DiagnosticPlots within dir if not provided an outdir.}

\item{...}{Additional parameters that can be fed to the palette? Not functional}
}
\description{
Give a directory that contains the output of an MFCL run and the function will automatically find the largest par file (or if is exists the out.par) to create all the figures currently available and save in a subdirectory called DiagnosticPlots
}
