% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.frqit.r
\name{plot.frqit}
\alias{plot.frqit}
\title{Plot a frqit::frq object}
\usage{
plot.frqit(Frq.list, Frq.names = NULL, fdesc = NULL, save.dir, save.name)
}
\arguments{
\item{Frq.list}{A list of Frq objects or a single Frq object. The reference model should be listed first.}

\item{Frq.names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the rep.list (if available) or generated automatically.}

\item{fdesc}{A data.frame with 7 columns (num,gear_long,method,code,gear,flag,region) and n rows, where n is the number of defined fisheries.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot a frqit::frq object
}
