% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kobe.r
\name{plot.kobe}
\alias{plot.kobe}
\title{Kobe plot}
\usage{
plot.kobe(
  rep,
  rep.name,
  agg.years = TRUE,
  agg.regions = TRUE,
  biomass.type = "SSB",
  save.dir,
  save.name
)
}
\arguments{
\item{rep}{An MFCLRep object.}

\item{rep.name}{A character string naming the models for plotting purposes}

\item{agg.years}{TRUE or FALSE. Should model outputs be aggregated to an annual time step.}

\item{agg.regions}{TRUE or FALSE. Should model outputs be aggregated across all regions are kept separate.}

\item{biomass.type}{Character string denoting the type of biomass plotted, 'SSB' or 'Total'}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Kobe plot
}
