% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.len.wt.temporal.r
\name{plot.len.wt.temporal}
\alias{plot.len.wt.temporal}
\title{plot.len.wt.temporal}
\usage{
plot.len.wt.temporal(
  par,
  rep,
  frq,
  max.wt = NULL,
  max.len = NULL,
  fdescloc = NULL,
  year.window = NULL,
  save.dir,
  save.name
)
}
\arguments{
\item{par}{MFCLPar object}

\item{rep}{MFCLRep object}

\item{frq}{MFCLFrq object}

\item{max.wt}{Optional: The maximum weight value to plot. Defaults to the maximum weight in the model}

\item{max.len}{Optional: The maximum length value to plot. Defaults to the maximum length in the model}

\item{fdescloc}{Optional: The path to the fdesc file to define the fishery names}

\item{year.window}{Optional: A single value giving the temporal window used to calculate the change in median length and weight from the beginning to the end of the model period. Defaults to 25% of the model period.}

\item{save.dir}{Optional: Path to the directory where the outputs will be saved}

\item{save.name}{Optional: Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the temporal change in median length and weight of vulnerable numbers by fishery
}
\details{
More documentation needed...
}
