% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.likelihood.profiles.R
\name{plot.likelihood.profiles}
\alias{plot.likelihood.profiles}
\title{Plot the likelihood profiles}
\usage{
plot.likelihood.profiles(llmodel, save.dir, save.name)
}
\arguments{
\item{llmodel}{A data.frame of profile information for one model.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the likelihood profiles using previously generated data for a single model.
The data generation is hidden away in other repos awaiting clean up.
It's based on Rob and Matt's code.
}
