% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.maturity.r
\name{plot.maturity}
\alias{plot.maturity}
\title{Plot maturity for each model give choice of by length or age}
\usage{
plot.maturity(
  pars,
  par.names = NULL,
  Length = FALSE,
  LnBins,
  show.legend = TRUE,
  palette.func = default.model.colours,
  xlab = "Age class",
  ylab = "Reproductive output",
  LegLoc = "bottomright",
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{pars}{a list of MFCLPar objects or a single MFCLPar object. The reference model should be the first in the list.}

\item{par.names}{an optional vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the par.list (if available) or generated automatically.}

\item{Length}{a boolean to determine whether to plot maturity at age (default) or maturity at length}

\item{LnBins}{a vector of length bin values}

\item{show.legend}{Do you want to show the plot legend, TRUE (default) or FALSE.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{xlab}{Name to display on the x axis label}

\item{ylab}{Name to display on the y axis label}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.colours to ensure consistency of model colours when plotting a subset of models.}

\item{LegLon}{legend location if show.legen is TRUE}
}
\description{
Plots the natural mortality by length bin or age class for each mode
}
