% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.movement.matrix.r
\name{plot.movement.matrix}
\alias{plot.movement.matrix}
\title{Plot the movement esimated movement for a single MFCLPar or the difference in the estimated movements for two pars with the same spatial, seasonal, age structure.}
\usage{
plot.movement.matrix(
  pars,
  par.names = NULL,
  age.vec = NULL,
  season.vec = NULL,
  save.dir,
  save.name
)
}
\arguments{
\item{pars}{a list of two MFCLPar objects or a single MFCLPar object. The reference model should be the first in the list.}

\item{par.names}{an optional vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the par.list (if available) or generated automatically.}

\item{age.vec}{An integer value or vector denoting the age class or classes to plot the movement for. If missing, the average movement across all ages is plotted.}

\item{season.vec}{An integer value or vector of integers denoting which seasons to plot the movement for. If missing then the average movement across seasons is plotted.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the movement esimated movement for a single MFCLPar or the difference in the estimated movements for two pars with the same spatial, seasonal, age structure.
}
