% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nat.mort.R
\name{plot.nat.mort}
\alias{plot.nat.mort}
\title{Plot natural mortality of each model}
\usage{
plot.nat.mort(
  rep.list,
  rep.names = NULL,
  show.legend = TRUE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{rep.list}{A list of MFCLRep objects or a single MFCLRep object. The reference model should be listed first.}

\item{rep.names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the rep.list (if available) or generated automatically.}

\item{show.legend}{Do you want to show the plot legend, TRUE (default) or FALSE.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.colours to ensure consistency of model colours when plotting a subset of models.}
}
\description{
Plots the natural mortality of each age class of each model
}
