% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.overall.composite.fit.R
\name{plot.overall.composition.fit}
\alias{plot.overall.composition.fit}
\title{Plot composite catch-at-length data for each fishery.}
\usage{
plot.overall.composition.fit(
  lfit,
  fisheries,
  fishery_names,
  save.dir,
  save.name,
  ncol = NULL,
  xlab = "Length (cm)"
)
}
\arguments{
\item{lfit}{A data.frame of observed and predicted composite catch-at-length data (generated using \code{length.fit.preparation()}).}

\item{fisheries}{The numbers of the fisheries to plot.}

\item{fishery_names}{The names of the fisheries to plot.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{ncol}{Number of columns to plot across. Default is ggplot2 default.}

\item{xlab}{Label for the xaxis.}
}
\description{
Plot composite (all time periods combined) observed and predicted catch-at-length for each fishery.
The table comes from the length.fit file and has already been processed using the
\code{length.fit.preparation()} function (which is based on Yukio's code).
This may also work with weight.fit data but has not been tested yet.
}
