% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pred.obs.catch.R
\name{plot.pred.obs.catch}
\alias{plot.pred.obs.catch}
\title{Plot difference between observed and predicted catches}
\usage{
plot.pred.obs.catch(
  rep.list,
  rep.names = NULL,
  fisheries,
  fishery_names,
  scale.diff = TRUE,
  show.legend = TRUE,
  show.points = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{rep.list}{A list of MFCLRep objects or a single MFCLRep object. The reference model should be listed first.}

\item{rep.names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the rep.list (if available) or generated automatically.}

\item{fisheries}{The numbers of the fisheries to plot.}

\item{fishery_names}{The names of the fisheries to plot.}

\item{show.legend}{Do you want to show the plot legend, TRUE (default) or FALSE.}

\item{show.points}{Do you want to show points as well as the smoother for the difference plots? Default is FALSE.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.colours to ensure consistency of model colours when plotting a subset of models.}
}
\description{
Plot difference between observed and predicted catches.
The differences can be scaled by the total catch of each fishery.
Multiple models can be passed in.
}
