% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pred.obs.cpue.R
\name{plot.pred.obs.cpue}
\alias{plot.pred.obs.cpue}
\title{Plot time series of the predicted and observed CPUE for selected fisheries}
\usage{
plot.pred.obs.cpue(
  frq = frq,
  par = par,
  fisheries,
  fishery_names,
  save.dir,
  save.name
)
}
\arguments{
\item{frq}{An object of type MFCLFrq that contains the observed catch, effort and penalty data.}

\item{par}{An object of MFCLPar that contains the effort devs.}

\item{fisheries}{The numbers of the fisheries to plot.}

\item{fishery_names}{The names of the fisheries to plot.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot observed (colored points) and model-predicted (black lines) CPUE for selected fisheries 
The color of the point indicates the penalty applied to the observed CPUE where brighter colors indicate a larger penalty.
}
\details{
The observed CPUE is calculated from the observed catches and effort in the 'freq' file.
The model-predicted CPUE is calculated from the observed catches in the 'freq' file and the model-estimated effort (from applying the estimated effort devs to the observed effort in the 'freq' file .)
The CPUE series are normalised by dividing by the mean before plotting.
(The difference between the predicted and observed time series is therefore just the effort devs).
}
