% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.plots.r
\name{plot.rec.devs}
\alias{plot.rec.devs}
\title{Plot estimated recruitment deviates}
\usage{
plot.rec.devs(par, save.dir, save.name)
}
\arguments{
\item{par}{An object of MFCLPar that contains the effort deviations.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot estimated recruitment deviates for a single model, by quarter and region.
A loess smoothed fit is shown.
}
