% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.selectivity.r
\name{plot.selectivity}
\alias{plot.selectivity}
\title{Compare the estimated selectivity across different models.}
\usage{
plot.selectivity(
  rep.list,
  rep.names = NULL,
  sel.basis = "AGE",
  palette.cols = c("royalblue3", "deepskyblue1", "gold", "orange1", "indianred1",
    "firebrick2", "#AC2020"),
  save.dir,
  save.name
)
}
\arguments{
\item{rep.list}{A list of MFCLRep objects or a single MFCLRep object. The reference model should be listed first.}

\item{rep.names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the rep.list (if available) or generated automatically.}

\item{sel.basis}{A character string indicating if selectivity at age ('AGE') or length ('Length') should be plotted}

\item{palette.cols}{A vector of character strings giving the colors to form a palette for differentiating between models. If wishing to use the exact colors give a vector of (length(rep.list) - 1) colors as the reference model is black by default.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Compare the estimated selectivity across different models.
}
