% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.attrition2.R
\name{Plot tag attrition}
\alias{Plot tag attrition}
\alias{plot.tag.attrition.v2}
\alias{prepare.tag.attrition.v2}
\alias{generate.plot.tag.attrition.v2}
\title{Plot the observed and predicted tag attrition.}
\usage{
plot.tag.attrition.v2(
  tagrelease.list,
  tagrep.list,
  fishery.map,
  names = NULL,
  facet = "program",
  plot.diff = TRUE,
  scale.diff = TRUE,
  show.legend = TRUE,
  show.points = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)

prepare.tag.attrition.v2(
  tagrelease.list,
  tagrep.list,
  names = NULL,
  fishery.map
)

generate.plot.tag.attrition.v2(
  pdat,
  facet = "program",
  model_names = unique(pdat$Model),
  plot.diff = TRUE,
  scale.diff = TRUE,
  show.legend = TRUE,
  show.points = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{tagrelease.list}{A list, or an individual data.frame, of tag release data created by the \code{tag.release.data.preparation()} function.}

\item{tagrep.list}{A list, or an individual data.frame, of tag returns from the temporary_tag_returns file. The output from the FLR4MFCL function \code{read.temporary_tag_report()}.}

\item{names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the tagrelease.list (if available) or generated automatically.}

\item{facet}{What variable fo you want to group by: "none" (no grouping), "program" (by tagging program - default), "region" (by recapture region).}

\item{plot.diff}{Do you want to plot the difference between the observed and predicted, or a time series of recaptures? TRUE (default) or FALSE.}

\item{scale.diff}{If TRUE, the difference between observed and predicted is scaled by the mean number of observed returns.}

\item{show.legend}{Do you want to show the plot legend, TRUE (default) or FALSE.}

\item{show.points}{Do you want to show points as well as the smoother for the difference plots? Default is FALSE.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.names to ensure consistency of model colours when plotting a subset of models.}

\item{pdat}{The output from calling prepare.plot.tag.returns.time.v2}

\item{model_names}{Names of the models to be plotted and ordered by.}
}
\description{
Plot the observed and predicted tag recaptures against time at liberty by tagging program, or all tagging programs combined.
The plot is either a time series of the difference between the observed and predicted, or a time series of the recaptures.
A loess smoother is put through the differences.
}
