% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.reporting.rate.R
\name{plot.tag.reporting.rate}
\alias{plot.tag.reporting.rate}
\title{Plot tag reporting rate}
\usage{
plot.tag.reporting.rate(
  par.list,
  model.names = NULL,
  reporting.rate.groups = NULL,
  plot_type = "violin",
  fill_colour = "lightblue",
  save.dir,
  save.name
)
}
\arguments{
\item{par.list}{A list of MFCLPar objects or a single MFCLFrq that contain the effort deviations.}

\item{model.names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the frqreal.list (if available) or generated automatically. Not really used here.}

\item{reporting.rate.groups}{A numerical vector of the reporting rate groups you want plotted. The vector can be named for more informative plot labels.}

\item{plot_type}{'violin' or 'box'.}

\item{fill_colour}{The fill colour.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
The tag reporting rates are reported by group. A fishery may belong to multiple groups, depending on the tag release event.
Plots a box or violing plot of the reporting rates for chosen reporting rate groups for all models in the input the list.
Note that this plot assumes the reporting rates groups are the same for each model in the list.
}
