% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.reporting.rates.r
\name{plot.tag.reporting.rates}
\alias{plot.tag.reporting.rates}
\title{Plot the tag reporting rates}
\usage{
plot.tag.reporting.rates(par, grp.names = NULL, save.dir, save.name)
}
\arguments{
\item{par}{An object of MFCLPar that contains the effort deviations.}

\item{grp.names}{Optional. Provide a character vector of names for the groups.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the tag reporting rates
}
