% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.returns.R
\name{plot.tag.return.proportion}
\alias{plot.tag.return.proportion}
\title{Plot proportions of tag returns by region}
\usage{
plot.tag.return.proportion(tagdat, plot.type = "point", save.dir, save.name)
}
\arguments{
\item{tagdat}{A data.frame of tagging data created by the \code{tag.data.preparation()} function.}

\item{plot.type}{What type of plot: 'point' (default) or 'bar'.}

\item{save.dir}{Path to the directory where the outputs will be saved.}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory.}
}
\description{
Plot the difference between the predicted and observed proportions of tag returns by region.
Experimental at the moment.
}
