% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.returns.prop.v2.R
\name{Plot tag return proportion}
\alias{Plot tag return proportion}
\alias{plot.tag.return.proportion.v2}
\alias{prepare.tag.return.proportion.v2}
\alias{generate.tag.return.proportion.v2}
\title{Plot proportions of tag returns by region}
\usage{
plot.tag.return.proportion.v2(
  tagrelease,
  tagrep,
  fishery.map,
  plot.type = "point",
  save.dir,
  save.name
)

prepare.tag.return.proportion.v2(tagrelease, tagrep, fishery.map)

generate.tag.return.proportion.v2(
  pdat,
  plot.type = "point",
  save.dir,
  save.name
)
}
\arguments{
\item{plot.type}{What type of plot: 'point' (default) or 'bar'.}

\item{save.dir}{Path to the directory where the outputs will be saved.}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory.}

\item{pdat}{Data from calling \code{prepare.tag.return.proportion.v2()}}

\item{tagdat}{A data.frame of tagging data created by the \code{tag.data.preparation()} function.}
}
\description{
Plot the difference between the predicted and observed proportions of tag returns by region.
Experimental at the moment.
}
