% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.returns.R
\name{plot.tag.returns.time}
\alias{plot.tag.returns.time}
\title{Plot the observed and predicted tag recaptures.}
\usage{
plot.tag.returns.time(
  tagrep,
  tagobs,
  par,
  fishery.map,
  plot.diff = TRUE,
  save.dir,
  save.name
)
}
\arguments{
\item{tagrep}{A data.frame of tag returns from the temporary_tag_returns file. The output from the FLR4MFCL function \code{read.temporary_tag_report()}.}

\item{tagobs}{An object of type MFCLTag.}

\item{par}{An object of MFCLPar that contains the effort deviations.}

\item{fishery.map}{A data.frame that describes which fishery is fishing in which region and with which gear. The columns are: fishery_name, region, gear, fishery (the reference number), tag_recapture_group and tag_recapture_name. Only tag groups in the fishery.map will be plotted.}

\item{plot.diff}{Do you want to plot the difference between the observed and predicted, or a time series of recaptures? TRUE (default) or FALSE.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the observed and predicted tag recaptures by tag recapture group.
Tags caught during the mixing period are not counted.
The plot is either a time series of the difference between the observed and predicted, or a time series of the recaptures.
A loess smoother is put through the differences.
}
