% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.data.preparation.R
\name{tag.data.preparation}
\alias{tag.data.preparation}
\title{Prepare data for tagging plots}
\usage{
tag.data.preparation(tagrep, tagobs, par, fishery.map)
}
\arguments{
\item{tagrep}{A data.frame of tag returns from the temporary_tag_returns file. The output from the FLR4MFCL function \code{read.temporary_tag_report()}.}

\item{tagobs}{An object of type MFCLTag.}

\item{par}{An object of MFCLPar that contains the effort deviations.}

\item{fishery.map}{A data.frame that describes which fishery is fishing in which region and with which gear. The columns are: fishery_name, region, gear, fishery (the reference number)and optionally tag_recapture_name, if not provided will default to the gear and region. The map should contain entries for all fisheries to avoid data being inadvertently dropped.}
}
\description{
The data required for the tagging plots are in different sources (temporary_tag_report, tag files, par files).
This function pulls the data together into a single data.frame to be passed to the tag plotting functions.
}
