% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.data.preparation.R
\name{tag.release.data.preparation}
\alias{tag.release.data.preparation}
\title{Prepare tag release data for tagging plots}
\usage{
tag.release.data.preparation(tagobs, par)
}
\arguments{
\item{tagobs}{An object of type MFCLTag.}

\item{par}{An object of MFCLPar that contains the effort deviations.}
}
\description{
The data required for the tagging plots are in different sources (temporary_tag_report, tag files, par files).
This function pulls the data together the release data from the tag file and the par file (for the mixing period).
This can be passed, along with the recapture data, into the tag plot functions
}
