% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.comparisons.r
\name{plot.comparisons}
\alias{plot.comparisons}
\title{Wrapper function for plotting model comparisons}
\usage{
plot.comparisons(
  rep.list,
  frq.list,
  par.list,
  model.names,
  agg.years = TRUE,
  agg.regions = TRUE,
  biomass.type = "SSB",
  biomass.units = 1000,
  fisheries.edev = NULL,
  recdist.year_range = NULL,
  yaxis.free = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{rep.list}{A list of MFCLRep objects. The reference model should be listed first.}

\item{frq.list}{A list of MFCLFrq objects}

\item{par.list}{A list of MFCLPar objects}

\item{model.names}{A vector of character strings naming the models for plotting purposes}

\item{agg.years}{TRUE or FALSE. Should model outputs be aggregated to an annual time step.}

\item{agg.regions}{TRUE or FALSE. Should model outputs be aggregated across all regions are kept separate.}

\item{biomass.type}{Character string denoting the type of biomass plotted, 'SSB' or 'Total'}

\item{biomass.units}{Integer number denoting how many MT to divide the biomass by. Default is 1000.}

\item{fisheries.edev}{A vector of fishery indices to plot for the effort devs}

\item{recdist.year_range}{The year range to average the total recruitment over. Default is the last 10 years.}

\item{yaxis.free}{TRUE or FALSE. Default is FALSE. If TRUE and agg.regions is also TRUE than the y-axis scales will be independent across regions, otherwise they will be shared so regional scaling will be apparent.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.colours to ensure consistency of model colours when plotting a subset of models.}
}
\description{
Wrapper function for plotting model comparisons
}
