% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.effort.devs.R
\name{plot.effort.devs}
\alias{plot.effort.devs}
\title{Plot effort deviations over time by fishery}
\usage{
plot.effort.devs(
  frq.list,
  par.list,
  fisheries,
  fishery.names,
  show.legend = TRUE,
  show.points = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{frq.list}{A named list MFCLFrq objects that contain the observed effort data.}

\item{par.list}{A named list of MFCLPar objects that contain the effort deviations.}

\item{fisheries}{The numbers of the fisheries to plot.}

\item{show.legend}{Do you want to show the plot legend, TRUE (default) or FALSE.}

\item{show.points}{Do you want to show points as well as the smoother for the difference plots? Default is FALSE.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.names to ensure consistency of model colours when plotting a subset of models.}

\item{fishery_names}{The names of the fisheries to plot.}
}
\description{
The effort deviations are in the 'par' object. The 'frq' object is needed to get the time series of fishing realisations.
Only effort deviations that have a corresponding effort in the 'frq' object are plotted.
A loess smoothed fit is shown.
}
