% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.effort.devs.R
\name{plot.effort.devs}
\alias{plot.effort.devs}
\title{Plot effort deviations over time by fishery}
\usage{
plot.effort.devs(frq, par, fisheries, fishery_names, save.dir, save.name)
}
\arguments{
\item{frq}{An object of type MFCLFrq that contains the observed effort data.}

\item{par}{An object of MFCLPar that contains the effort deviations.}

\item{fisheries}{The numbers of the fisheries to plot.}

\item{fishery_names}{The names of the fisheries to plot.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
The effort deviations are in the 'par' object. The 'frq' object is needed to get the time series of fishing realisations.
Only effort deviations that have a corresponding effort in the 'frq' object are plotted.
A loess smoothed fit is shown.
}
