% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.F.temporal.r
\name{plot.F.temporal}
\alias{plot.F.temporal}
\title{Compare Fishing Mortality Across Different Models}
\usage{
plot.F.temporal(
  x,
  par.list = NULL,
  rep.names = NULL,
  agg.years = TRUE,
  agg.regions = TRUE,
  agg.ages = NULL,
  yaxis.free = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{x}{A list of \code{MFCLRep} objects or a single \code{MFCLRep} object.
The reference model should be listed first.}

\item{par.list}{Optional list of \code{MFCLPar} objects or a single
\code{MFCLPar} object. Used for plotting juvenile and adult fishing
mortality. If specified then \code{agg.ages} is ignored.}

\item{rep.names}{A vector of character strings naming the models for plotting
purposes. If not supplied, model names will be taken from the names in
the \code{x} list (if available) or generated automatically.}

\item{agg.years}{\code{TRUE} or \code{FALSE}. Should model outputs be
aggregated to an annual time step.}

\item{agg.regions}{\code{TRUE} or \code{FALSE}. Should model outputs be
aggregated across all regions or kept separate.}

\item{agg.ages}{A vector of age classes to average F over. Default is all age
classes.}

\item{yaxis.free}{\code{TRUE} or \code{FALSE}. If \code{TRUE} and
\code{agg.regions} is also \code{TRUE} then the y-axis scales will be
independent across regions, otherwise they will be shared to show
regional scaling.}

\item{palette.func}{A function to determine the colours of the models. The
default palette has the reference model in black. It is possible to
determine your own palette function. Two functions currently exist:
\code{default.model.colours} and \code{colourblind.model.colours}.}

\item{save.dir}{Path to the directory where the outputs will be saved.}

\item{save.name}{Name stem for the output, useful when saving many model
outputs in the same directory.}

\item{\dots}{Extra arguments passed to \code{palette.func}. Use the argument
\code{all.model.colours} to ensure consistency of model colours when
plotting a subset of models.}
}
\description{
Compare Fishing Mortality Across Different Models
}
