% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.plots.r
\name{plot.rec.devs}
\alias{plot.rec.devs}
\title{Plot estimated recruitment deviates}
\usage{
plot.rec.devs(
  par.list,
  par.names = NULL,
  show.legend = TRUE,
  show.points = FALSE,
  palette.func = default.model.colours,
  save.dir,
  save.name,
  ...
)
}
\arguments{
\item{par.list}{A list of MFCLRep objects or a single MFCLRep object. The reference model should be listed first.}

\item{par.names}{A vector of character strings naming the models for plotting purposes. If not supplied, model names will be taken from the names in the rep.list (if available) or generated automatically.}

\item{show.legend}{Do you want to show the plot legend, TRUE (default) or FALSE.}

\item{show.points}{Do you want to show points as well as the smoother for the difference plots? Default is FALSE.}

\item{palette.func}{A function to determine the colours of the models. The default palette has the reference model in black. It is possible to determine your own palette function. Two functions currently exist: default.model.colours() and colourblind.model.colours().}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}

\item{...}{Passes extra arguments to the palette function. Use the argument all.model.names to ensure consistency of model colours when plotting a subset of models.}
}
\description{
Plot estimated recruitment deviates for a single model, by quarter and region.
A loess smoothed fit is shown.
}
