% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.returns.R
\name{plot.tag.attrition}
\alias{plot.tag.attrition}
\title{Plot the observed and predicted tag attrition.}
\usage{
plot.tag.attrition(
  tagdat,
  facet.program = TRUE,
  plot.diff = TRUE,
  scale.diff = TRUE,
  save.dir,
  save.name
)
}
\arguments{
\item{tagdat}{Tagging data created by the \code{tag.data.preparation()} function.}

\item{facet.program}{Do you want to process and plot by tagging program, or combine the tagging programs. TRUE (default) or FALSE.}

\item{plot.diff}{Do you want to plot the difference between the observed and predicted, or a time series of recaptures? TRUE (default) or FALSE.}

\item{scale.diff}{If TRUE, the difference between observed and predicted is scaled by the number of observed returns.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the observed and predicted tag recaptures against time at liberty by tagging program, or all tagging programs combined.
The plot is either a time series of the difference between the observed and predicted, or a time series of the recaptures.
A loess smoother is put through the differences.
}
