% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.returns.R
\name{plot.tag.returns.time}
\alias{plot.tag.returns.time}
\title{Plot the observed and predicted tag recaptures.}
\usage{
plot.tag.returns.time(
  tagdat,
  recapture.groups,
  plot.diff = TRUE,
  scale.diff = TRUE,
  save.dir,
  save.name
)
}
\arguments{
\item{tagdat}{Tagging data created by the \code{tag.data.preparation()} function.}

\item{recapture.groups}{A vector of the reference numbers of the tag recapture groups you want to plot.}

\item{plot.diff}{Do you want to plot the difference between the observed and predicted, or a time series of recaptures? TRUE (default) or FALSE.}

\item{scale.diff}{If TRUE, the difference between observed and predicted is scaled by the number of observed returns.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot the observed and predicted tag recaptures by tag recapture group.
Tags caught during the mixing period are not counted.
The plot is either a time series of the difference between the observed and predicted, or a time series of the recaptures.
A loess smoother is put through the differences.
}
